% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_make.R
\name{parse_makefile}
\alias{parse_makefile}
\title{Extract targets and dependencies from a Makefile}
\usage{
parse_makefile(makefile = "Makefile")
}
\arguments{
\item{makefile}{A filename or file connection specifying the Makefile (Default: "Makefile")}
}
\value{
a tibble
}
\description{
The function tries to identify all rules contained in a Makefile,
returning a tidy data frame with a mapping between targets and
dependencies. Each rule from the Makefile is thus transformed
to give one row per dependency. The original Makefile rules are
preserved in the columns ´rule_id` and `rule`.
}
\details{
No attempt is made to fully parse the Makefile according to the
GNU make specification. In particular, the following elements are
ignored: imported files, commands, comments, variables.
}
