% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_makefile.R
\name{plot_makefile}
\alias{plot_makefile}
\title{Plot Makefile rules as a graph}
\usage{
plot_makefile(m)
}
\arguments{
\item{m}{A parsed Makefile, obtained from `parse_makefile`}
}
\description{
Plot Makefile rules as a graph
}
