% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_make.R
\name{use_make}
\alias{use_make}
\title{Copy a Makefile template to project root directory}
\usage{
use_make()
}
\value{
Output of file.copy, signifying success of operation
}
\description{
This function is inspired by the `usethis` package. It copies a Makefile
template to the project root, containing generic rules for building R
projects.
}
